requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Deep Learning
tensorflow>=2.10.0
torch>=1.12.0
torchvision>=0.13.0

# Vision Transformers
transformers>=4.20.0
timm>=0.6.0

# Image processing
opencv-python>=4.6.0
Pillow>=9.0.0
albumentations>=1.3.0

# Explainability
lime>=0.2.0.1
grad-cam>=1.4.0
tf-keras-vis>=0.8.0

# AWS
boto3>=1.24.0
sagemaker>=2.100.0
awscli>=1.25.0

# Data download
gdown>=4.5.0
google-api-python-client>=2.50.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0

# MLOps
mlflow>=2.0.0
dvc>=2.10.0
wandb>=0.13.0

# Configuration
pyyaml>=6.0
python-dotenv>=0.20.0

# Utilities

tqdm>=4.64.0
requests>=2.28.0
pydantic>=1.9.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.9.0

# API
fastapi>=0.95.0
uvicorn>=0.20.0
python-multipart>=0.0.5

# Testing
pytest>=7.1.0
pytest-cov>=3.0.0
pytest-mock>=3.8.0

# Code quality
black>=22.6.0
flake8>=4.0.0
mypy>=0.961
isort>=5.10.0

# Logging and monitoring
python-json-logger>=2.0.0
prometheus-client>=0.14.0

# Data validation
great-expectations>=0.15.0
evidently>=0.2.0

# Jupyter
jupyter>=1.0.0
notebook>=6.4.0
ipywidgets>=7.7.0